Improve some type definitions and decrease specialization
This PR is an alternative to #310, although it would be possible to
add precompiles to this in a later PR. In general, I've found that the
first and most impactful step is to analyze specialization costs using
SnoopCompile's `pgdsgui` tool. Another reason to do this first is
because it affects the MethodInstances and hence the specific
precompiles that get generated.

Using the following benchmark,
```julia
@time begin
    include("test/rw.jl")
    include("test/loadsave.jl")
end
```
I get the following timings:

master prior to merge of #310: 65.2s
master: 61.6s (so yes, it helped)
this PR: 47.4s