convert FileIO and UnPack to true extensions
Avoids the recursion of having FileIO expect to load JLD2 in the
precompile script, triggering possible errors. The functions load and
save were not previously documented as part of the API (although they
still exist here as public functions when desired), and the tests even
usually explicitly got them from FileIO when desired.